Класс для тестирования DatabaseHandler. Первый тест

+++ lib/src/main/java/org/company/cardstat/DatabaseHandler.java
@@ -302,7 +302,7 @@ public class DatabaseHandler extends SQLiteOpenHelper {
      * @return
      * @throws DatabaseHandlerException
      */
-    public BankMessage getBankMessage(int _id) throws DatabaseHandlerException {
+    public BankMessage getBankMessage(long _id) throws DatabaseHandlerException {
 
         SQLiteDatabase database = this.getReadableDatabase();
